test: refactor calibration feature tests to match single file feature layout and add new ion coverage, sequence and token feature tests
